requirements.txt
fastapi
httpx
openai
pydantic
pydantic_settings
requests
starlette
uvicorn
google-genai
jinja2
python-multipart
cryptography # 支持 MySQL 8+ caching_sha2_password 验证
# 数据库相关依赖
pymysql
sqlalchemy
aiomysql
databases
python-dotenv
apscheduler # 添加定时任务库

packaging
